Add a juju charm (#10)
This is based almost entirely on the charm in javan-rhino, with minor
tweaks.  build.snapcraft.io makes a terrible charm name, so (pending
bikeshedding) I've gone for "snap-build" by analogy with a few other OLS
snap-* services.

memcached is included even though we aren't using it yet, since it seems
probable we'll find some use for it.