Make development configuration charm-friendly (#14)
* Reload systemd configuration after rendering unit

Otherwise systemd isn't guaranteed to do the right thing on restart.

* Change development listen ports to 8000/8001

This makes it easier to make the charm work on all environments (since
staging and production will use 8000).

* Make development server listen on 0.0.0.0

Otherwise it's effectively impossible to deploy a test development
server using juju.  It does mean you have to be a bit careful that you
have a firewall in the way if you're standing up a test service
somewhere, but that was surely the case for other reasons anyway.