Drop /api from Launchpad service root (#38)
The client in the Launchpad source tree uses /api/devel as its service
root because it's running on launchpad.net itself, and in that context
it's easier to use the webservice on https://launchpad.net/api/devel/
and not have to worry about cross-origin issues.

However, in our case we're running on a different domain and using
api.launchpad.net, and https://api.launchpad.net/api/devel/ seems to
have at least some of the same rules as https://launchpad.net/api/devel/
in that it requires a Referer header.  That's awkward and unnecessary
here.  Using just https://api.launchpad.net/devel/ is simpler and avoids
this requirement.