Use signed webhooks (#54)
Webhooks need to be signed in order to (e.g.) avoid easy DoSing of the
build farm without authentication.  Set up a root webhook secret which
is HMACed with the account name and repository name to create a
per-repository secret.  This requires moving the notification endpoint
to `/:account/:repo/webhook/notify` so that we can easily extract the
necessary information to verify the signature.

If the root secret ever needs to change, we'll be in a bit of trouble
since updating it requires GitHub credentials from the user, but that's
probably unavoidable and if necessary we can probably deal with it
somehow once the user logs into build.snapcraft.io.