Fix spurious promise rejection in internalFindSnap (#81)
If internalFindSnap doesn't find the given repository URL in memcached,
it looks for it in Launchpad, iterating through all snaps with that URL.
If one of them matches the service username, it stores that in memcached
and resolves the promise with its link; otherwise, it rejects the
promise with a snap-not-found error.

However, this code incorrectly fell through to the last-resort promise
rejection, and because memcached set is asynchronous the rejection would
typically be processed before the resolution.  We didn't catch this
because the memcached stub in tests was synchronous.  Avoid this
incorrect fallthrough and strengthen the tests to catch bugs of this
form.

Fixes #79.