Check snapcraft.yaml existence in webhook (#109)
We want to move to creating the snap in Launchpad somewhat earlier,
before snapcraft.yaml necessarily exists.  As a prerequisite for this,
the webhook needs to check whether snapcraft.yaml exists before
dispatching builds.

The `auto_build` flag on snaps, previously unused in this workflow, is
now used to indicate whether the snap has been sufficiently set up for
builds to be dispatched.  This lets us minimise the number of GitHub
queries we make from the webhook, which is important since that doesn't
have user authorisation available to it.