Initiate the snapcraft.yaml check from the client (#114)
We'll soon need to move the process of getting a package_upload macaroon
to the client side, which means that the client will need to know the
snap name when it's creating a snap.  Add an
`/api/github/snapcraft-yaml/:owner/:name` route to fetch the parsed
file, and change `/api/launchpad/snaps` to take the snap name, series,
and channels as parameters; this lets us control more of the data flow
from the client side.