Add actions for checking store sign-in (#126)
We need to be able to check whether the user is already signed into the
store; the existing actions only arrange to set authStore.authenticated
on the first page load.  These new actions check local storage for
existing macaroons.

We should check the expires caveat on existing macaroons; this is
compounded by us not yet clearing local storage when the user signs out
(#125).  In the short term a workaround is to delete entries from
storage manually using devtools.