Use package_upload_request macaroons for authz (#127)
Creating snaps now requires being "signed into the store", which equates
to having a discharged package_upload_request macaroon in local storage.
For each snap to be created, we request a specific non-expiring
package_upload macaroon for it and tell the backend to store that in
Launchpad.  This allows batch-creation of snaps, and avoids a round-trip
through SSO per snap.

This leaves some unused endpoints around, but this PR is already large
enough; they can be cleaned up later.