Create webhook before creating snap (#206)
We used to create a webhook during the setup workflow, but that fell out
during the major reorganisations of the last couple of weeks.  This
restores it.  We can safely create the webhook as the very first thing
we do, since the webhook handler will fail cleanly if the snap doesn't
exist yet, and error handling is easier this way round.

This duplicates part of `src/common/actions/webhook.js`, but that can
just be removed after this is in place; we don't need to have separate
actions and state for this.