Trigger builds if possible once name is registered (#202)
Once snapcraft.yaml exists and the name is registered, we can trigger
builds.  One path for this (already implemented) is that if the name is
registered first then we may be told of a new snapcraft.yaml by way of a
webhook push event.  The other possibility is implemented here: if
snapcraft.yaml already exists when the name is registered, we should
trigger builds at that point.