Only set auto_build once a snap is fully set up (#258)
BSI mainly uses the `auto_build` flag on a snap in Launchpad as a signal
to the webhook handler that it should dispatch builds.  However, it's
also piggy-backing on a Launchpad feature that dispatches builds for
"stale" snaps: the exact details of this are mostly not relevant for
BSI, but one relevant point is that a new snap is considered stale until
the first time `requestAutoBuilds` is called on it.  Setting
`auto_build` on a new snap therefore risks the dispatch of duplicate
builds depending on the exact timings involved.

We now set `auto_build` only when the snap has a `snapcraft.yaml` and
its name is registered, and we do so immediately after calling
`requestAutoBuilds` so that duplicate builds won't happen.

Fixes #225.