Convert internal LP client to async/await (#376)
This is a small proof-of-concept to see if people like this before I
start sending PRs to convert the whole codebase.  I find this style
easier to read and less error-prone, especially around error handling,
and it generally decreases indentation levels.

I'm deliberately not converting tests at the same time as the code under
test, to provide better assurance that I haven't applied a mistaken
conversion to both.