Add a /metrics endpoint (#419)
For now, this just exposes the basic set of metrics provided by
`express-prom-bundle`.  In future we'll add our own as well.

We now pay attention to the `trusted_networks` charm configuration
option provided by `layer:ols`.  On staging and production, that will be
used to restrict access to the `/metrics` endpoint.

Part of #360.