Tighten up reactive state handling slightly (#461)
`configure` should only run when the `ols.pg.configured` state is set;
otherwise it's possible for this function to run, set the unit's status
to "active", and then for `ols_pg_configured` to run immediately
afterwards and set the status back to "waiting".  This is usually
harmless since hook retries will cause things to settle down eventually,
but it can slow down deployments by several minutes.