Point Prometheus at public frontends (#481)
Monitoring the individual appservers turns out to be awkward because
we'd have to expose the relevant port in security groups.  We don't
actually care about the per-unit monitoring anyway, only the
application-wide metrics that come from the database and can be
published by any appserver, so let's just point the monitoring server at
the frontend instead.

This requires the charm to know the application's base URL.  For this,
I've taken a dual approach: firstly, I've added a hack to fish it out of
the environment file for now, to minimise the number of steps needed to
get this deployed; and secondly, I've added a `base_url` charm
configuration option, which fits with the general goal of eventually
superseding the `.env` files.

Part of #360.