Refactor redux part i (#541)
* refactor repositories state, fixes #464

* fix add-repos spinner post rebase

* Fix pageSlice failing when there are no pageLinks (#574)

* Whitelist and camelCase props from GH API (#578)

* remove typo/skipped tests causing parse failure (#579)

* Updated action names for adding repo (#580)

* Better select-repository-list component test coverage (#581)

* more test coverage for select-repositories-list

* update for action name changes

* use length from fixture

* use new action creator names (#584)

* use new action creator names

* Update t_repositories.js

* skip failing tests, temporarily (#585)

* skip failing tests, temporarily

* skip failing metrics test

* silence

* Cleaning up entities reducer code (#587)

* fix select-repository-row tests (#588)

* fix success response fixture (#589)

* Drop __ prefix from internal entities props (#590)

* shonkwrap from fresh install (#594)

* add comment on normalizr use in lp handler (#593)

* add comment on normalizr use in lp handler

* Update launchpad.js

* update selector to use new state (#592)

* More test coverage for repository actions and reducers (#597)

* entities default test

* repository entity tests

* Update t_entities.js

* Update t_entities.js

* keep entity collection reducers together

* use repoPayload fixture

* Merge master into refactor redux part i (#598)

* 495: CSRF tokens (#553)

* 495: CSRF tokens

- Added new Express middleware to generate and verify CSRF tokens
- Added global window.__CSRF_TOKEN__ variable for application to consume
  token from server
- Added new API_CALL redux middleware to encapsulate API requests and pass tokens
- Migrated FETCH_SNAPS and REMOVE_SNAPS chain of events to API_CALL
- Fixed snaps reducer to work with API_CALL middleware
- Added unit tests

* 495: Remove TODO

* 495: CSRF Tokens

- Removed redux-mock-store from snaps unit tests
- Removed thunks from snaps actions

* 495: CSRF tokens

- Revised payload shape for CALL_API middleware

* 495: Added test for callApi middleware

* 495: CSRF Tokens

- Rearranged some server routes
- Removed a duplicate route

* Add metric for builds released to edge (#566)

* Add metric for builds released to edge

We do this by configuring a webhook on Launchpad snaps and counting
events where `store_upload_status` is "Uploaded".  (We could use the
information provided by this webhook for some more things in future to
reduce polling, although that would also require putting some more
thought into ordering guarantees; since "Uploaded" is a terminal status
we don't need to think about that too hard in this case.)

Of course there are lots of existing snaps that need to have the webhook
added, and it's also possible that we might create the snap and then
fail to create the webhook.  We don't currently have any infrastructure
for running scripts or periodic jobs, which would be a better way to
handle this; but as a reasonably lightweight workaround for these
problems, `/api/launchpad/snaps/list` also ensures that the necessary
webhook exists for each of the snaps it returns, and we make a note in
memcached that we've done so to avoid hitting the Launchpad API too
often.

Fixes #517.

* Break up webhook handler into separate functions

* Document lp:has_webhook:<snap>

* describe → context

* Replace node uuid (#577)

* replace-node-uuid

- Updated shrinkwrap against master

* replace node-uuid

- Installed uuid module
- Uninstalled node-uuid

* Don't generate CSRF token if there is no session (#576)

* Log information for register-name 401 failures (#575)

* Move getCaveats to a separate helper module

* Log information for register-name 401 failures

This should help to debug #527.

* update-react (#547)

- Updated react and react-dom to 15.4.2

* Adjust the alignment of the beta flag (#573)

* Drop next-route fallthrough from universal handler (#586)

This has recently started causing `HEAD /` requests to fail on staging.
I'm not clear on why this has only started happening recently, but I
think this code has been wrong for some time: `handleMatch` is always
called and emits a response in all cases, so there's no reason to fall
through to other routes here (except for error-handling middleware,
which will still be called on unhandled exceptions).

Fixes #583.

* 30 seconds beta notification delay (#591)

* New tag manager container id for tracking (#595)

* Fixing passing snaps to entities and ids

* Cleaning up some merge comments

* drop union (#601)

* Merge master (#602)

* 495: CSRF tokens (#553)

* 495: CSRF tokens

- Added new Express middleware to generate and verify CSRF tokens
- Added global window.__CSRF_TOKEN__ variable for application to consume
  token from server
- Added new API_CALL redux middleware to encapsulate API requests and pass tokens
- Migrated FETCH_SNAPS and REMOVE_SNAPS chain of events to API_CALL
- Fixed snaps reducer to work with API_CALL middleware
- Added unit tests

* 495: Remove TODO

* 495: CSRF Tokens

- Removed redux-mock-store from snaps unit tests
- Removed thunks from snaps actions

* 495: CSRF tokens

- Revised payload shape for CALL_API middleware

* 495: Added test for callApi middleware

* 495: CSRF Tokens

- Rearranged some server routes
- Removed a duplicate route

* Add metric for builds released to edge (#566)

* Add metric for builds released to edge

We do this by configuring a webhook on Launchpad snaps and counting
events where `store_upload_status` is "Uploaded".  (We could use the
information provided by this webhook for some more things in future to
reduce polling, although that would also require putting some more
thought into ordering guarantees; since "Uploaded" is a terminal status
we don't need to think about that too hard in this case.)

Of course there are lots of existing snaps that need to have the webhook
added, and it's also possible that we might create the snap and then
fail to create the webhook.  We don't currently have any infrastructure
for running scripts or periodic jobs, which would be a better way to
handle this; but as a reasonably lightweight workaround for these
problems, `/api/launchpad/snaps/list` also ensures that the necessary
webhook exists for each of the snaps it returns, and we make a note in
memcached that we've done so to avoid hitting the Launchpad API too
often.

Fixes #517.

* Break up webhook handler into separate functions

* Document lp:has_webhook:<snap>

* describe → context

* Replace node uuid (#577)

* replace-node-uuid

- Updated shrinkwrap against master

* replace node-uuid

- Installed uuid module
- Uninstalled node-uuid

* Don't generate CSRF token if there is no session (#576)

* Log information for register-name 401 failures (#575)

* Move getCaveats to a separate helper module

* Log information for register-name 401 failures

This should help to debug #527.

* update-react (#547)

- Updated react and react-dom to 15.4.2

* Adjust the alignment of the beta flag (#573)

* Drop next-route fallthrough from universal handler (#586)

This has recently started causing `HEAD /` requests to fail on staging.
I'm not clear on why this has only started happening recently, but I
think this code has been wrong for some time: `handleMatch` is always
called and emits a response in all cases, so there's no reason to fall
through to other routes here (except for error-handling middleware,
which will still be called on unhandled exceptions).

Fixes #583.

* 30 seconds beta notification delay (#591)

* New tag manager container id for tracking (#595)

* keep selected state on success (#603)

* keep selected state on success

* Update repository.js