Add a trivial /_status/check endpoint (#1019)
This is just enough to keep the ols-http charm layer happy.  I'm not
quite sure why this suddenly started to be a problem, but it's currently
causing deployments to fail to complete.

We don't have a good way to unit-test the full server app assembly, but
I've tested this manually by pointing curl at `/_status/check`, either
from localhost or from an IP address in `TRUSTED_NETWORKS`.