Move "Build trigger" column before "Architecture"
The "Build trigger" column is currently missing from the design
specification.  Design work for #556 calls for failed build requests to
have their request ID and error message rendered spanning the
"Architecture", "Duration", and "Result" columns; since failed build
requests still have a meaningful build trigger, having "Build trigger"
in the middle of this group of columns makes it hard to comply with that
otherwise-reasonable suggestion.

Moving "Build trigger" two columns to the left resolves this difficulty,
and seems to make just as much sense.  In fact, once we start using
build requests, it arguably makes more sense (because the build trigger
is known before the set of built-on architectures is known).