Prepare UI for upcoming build request changes (#1152)
We will shortly need to change new snap build requests to be dispatched
more asynchronously, in order that Launchpad can fetch `snapcraft.yaml`
from the repository and decide which architectures the snap should be
built for.  This means that a snap can be in a state where it has
pending build requests (different from pending builds: for instance,
they don't have an architecture), and also where it has failed build
requests in its history (for example, because `snapcraft.yaml` is
malformed in such a way that Launchpad can't determine which
architectures it should be built for).

This prepares the UI layer for that work by ensuring that it can
understand the modified server responses and render build requests in
the builds list.  The server side hasn't been changed yet (that will be
a follow-up PR), so this should result in no functional change for the
moment.

Build requests will be interleaved with builds in server responses, and
the UI tells them apart by looking at the `resource_type_link`.  This
wasn't my first choice for designing the internal API, but it turns out
to be the simplest option: in particular, if we want to be able to
support sensible pagination of the combined build and build request
history in future then we need to be iterating over a single collection.

Part of #556.