Fetch snapcraft.yaml before requesting builds (#1162)
It's possible for a repository to have once been configured on
build.snapcraft.io, but for the contributors to that repository to later
remove snapcraft.yaml for whatever reason.  We have a few cases of this
on production today, and they generate a large volume of email
notifications to the operators of the ~build.snapcraft.io Launchpad
account because there ends up being a failed build request for every
commit to one of the repositories in question.

The simplest way to avoid this is to always check for a snapcraft.yaml
before dispatching automatic builds of a snap (webhook or poller).  This
does have the downside that the webhook won't automatically cause failed
build requests if snapcraft.yaml is present but syntactically invalid;
but if somebody is trying to debug that then they'll probably try to
dispatch a manual build, and the problem will become clear at that
point, at least once we start displaying recent failed build requests.

While I was here, I took the opportunity to make some of the poller
tests a bit more precise about expected URLs.